fixed join requests
