+++ client/src/app/dashboard/settings/project/ActiveInvites.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+const ActiveInvites = (props: Props) => {
+  return (
+    <div>ActiveInvites</div>
+  )
+}
+
+export default ActiveInvites
