+++ client/src/app/dashboard/settings/project/CriticalSettings.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+type Props = {}
+
+const CriticalSettings = (props: Props) => {
+  return (
+    <div>CriticalSettings</div>
+  )
+}
+
+export default CriticalSettings
